requirements.txt
seaborn
catboost
xgboost
lightgbm
django-emoji
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.6
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pandas-stubs==2.3.2.250827
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.9.1
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
wordcloud==1.9.4
